refactor(webrtc): Extract reusable components into focused modules
Major architectural refactoring to improve code organization, reusability, and maintainability.

## New Module Structure

**Core Streaming Utilities** (cross-platform):
- `core/streaming/opus.rs` (195 lines) - Opus audio encoding
- `core/streaming/rtp.rs` (150 lines) - RTP conversion and timestamp calculation

**Apple WebRTC Infrastructure** (platform-specific utilities):
- `apple/webrtc/whip.rs` (416 lines) - WHIP HTTP signaling (RFC 9725)
- `apple/webrtc/session.rs` (774 lines) - WebRTC session management

**VideoToolbox Enhancements**:
- Moved H.264 NAL parsing from rtp.rs to videotoolbox/format.rs (better semantic fit)
- Added public exports for NAL unit parsing utilities

## Impact

**processors/webrtc.rs reduced from 2599 → 913 lines (65% reduction)**

- Removed 1686 lines of reusable components
- Processor now focuses solely on pipeline integration
- All infrastructure code properly separated into focused modules

## Benefits

1. **Reusability**: OpusEncoder, RTP utilities, WHIP client, and WebRTC session can be used independently
2. **Maintainability**: Each module has clear responsibility, manageable size (150-800 lines)
3. **Testability**: Components can be tested in isolation with clean interfaces
4. **Architecture**: Clear separation of processors vs infrastructure/utilities

## Components Extracted

- **OpusEncoder**: Platform-agnostic Opus encoding (can be used by MP4 writer, RTMP, etc.)
- **RTP utilities**: Generic RTP timestamp calculation and sample conversion
- **WhipClient**: Reusable WHIP signaling client (RFC 9725 compliant)
- **WebRtcSession**: Complete WebRTC session management (ICE, DTLS, tracks, SRTP)
- **NAL parsing**: H.264 AVCC/Annex B NAL unit parsing (now in videotoolbox module)

## Zero Functionality Changes

- Pure refactoring with no logic changes
- All existing APIs preserved
- WebRTC streaming verified working with Cloudflare Stream
- Build successful with no new errors
