+++ libs/streamlib/src/apple/mod.rs
@@ -8,6 +8,7 @@ pub mod metal;
 pub mod texture;
 pub mod wgpu_bridge;
 pub mod videotoolbox;
+pub mod webrtc;
 
 pub mod processors;
 
@@ -37,12 +38,13 @@ pub use processors::{
     AppleDisplayProcessor,
     AppleAudioOutputProcessor,
     AppleMp4WriterProcessor,
-    // WebRTC WHIP processor and config types:
+    // WebRTC WHIP processor:
     WebRtcWhipProcessor,
     WebRtcWhipConfig,
-    WhipConfig,
-    AudioEncoderConfig,
 };
+
+// Re-export webrtc types
+pub use webrtc::{WhipClient, WhipConfig, WebRtcSession};
 
 // Re-export videotoolbox types (VideoEncoderConfig and H264Profile now come from videotoolbox module)
 pub use videotoolbox::{VideoCodec, H264Profile, VideoEncoderConfig, VideoToolboxEncoder};
