+++ libs/streamlib/src/apple/webrtc/mod.rs
@@ -1,0 +1,9 @@
+// WebRTC Implementation for macOS/iOS
+//
+// Provides WHIP signaling and WebRTC session management.
+
+pub mod whip;
+pub mod session;
+
+pub use whip::{WhipClient, WhipConfig};
+pub use session::WebRtcSession;
