+++ libs/streamlib/src/core/streaming/mod.rs
@@ -1,0 +1,9 @@
+// Streaming Utilities
+//
+// Core streaming components for audio/video encoding and RTP/WebRTC.
+
+pub mod opus;
+pub mod rtp;
+
+pub use opus::{OpusEncoder, AudioEncoderOpus, AudioEncoderConfig, EncodedAudioFrame};
+pub use rtp::{convert_video_to_samples, convert_audio_to_sample, RtpTimestampCalculator};
